refactor: split out contract test runner
